requirements.txt
aiohttp
colorama
discord.py
markdownify
pytz
Requests
pyutils @ git+https://github.com/hutnerr/python-utils.git

